mgmt/glusterd: fix warnings during build.

BUG: 1307 (gluster volume defrag <VOLNAME> status)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=1307

+++ xlators/mgmt/glusterd/src/glusterd-handler.c
@@ -662,8 +662,9 @@ glusterd_check_and_rebalance (glusterd_volinfo_t *volinfo, char *dir)
                 }
                 if (S_ISREG (stbuf.st_mode) && ((stbuf.st_mode & 01000) == 01000)) {
                         /* TODO: run the defrag */
-                        snprintf (tmp_filename, 1024, "%s/.%s.gfs%zu", dir,
-                                  entry->d_name, stbuf.st_size);
+                        snprintf (tmp_filename, 1024, "%s/.%s.gfs%llu", dir,
+                                  entry->d_name,
+                                  (unsigned long long)stbuf.st_size);
 
                         dst_fd = creat (tmp_filename, (stbuf.st_mode & ~01000));
                         if (dst_fd == -1)
